Test suite: allow the pattern to cover slightly less of the write cache
I've seen a few cases when it covered "only" 93.5%. IMO that's good enough.

Similarly to when I did 03a0f091461620033dbbf8344d95954877ddfd89, I suppose
something is happening with this newer kernel, which can occasionally make the
kernel reclaim that some of write cache. Or, the kernel already started to write
this data out to disk, and thus removed it from its write cache.