Tor Connection: Fix handling of case that image can't be found
Currently, when the image shown in tca
(/usr/share/doc/tails/website/about/relays.svg) can't be found, an
exception is thrown and the initialization of the main window is
aborted. This leaves the main window in the "Connect to a local network"
state, even when a local network connection already exists.

This commit logs the exception instead and continues with the
initialization.

This issue is only relevant for development purposes, where the image
file might not exist.