Upgrader: don't crash when a download is interrupted and then resumed
When a download is interrupted and then resumed due to a failure or a shutdown
in the network connection, results in bytes_downloaded=0. When this happens, the
download speed could get a negative value which causes an error because negative
numbers are not allowed in format_bytes.

This merely avoids the worst consequence of this bug, namely: it makes it
impossible to upgrade. There's more to it on the issue.

refs #18435