tps: Treat unmounted partition as locked
We know of two cases which causes the partition to be unlocked but not
mounted:

* User did some operations manually, e.g. via Disks (unlocked but not
  mounted or unmounted a previously unlocked partition).
* Unlocked successfully in the Welcome Screen but failed to mount with
  ext4 errors.

Previously, the frontend allowed users to try activate/deactivate
features when the partition was unlocked but not mounted. This resulted
in unhelpful error messages.

To avoid that, we now treat a partition in this state as locked. In this
state, the frontend currently tells the user:

    To use or configure your Persistent Storage, unlock it in the
    Welcome Screen when starting Tails.

Note that an Unlock call to the backend when in this state does correctly
mount the Persistent Storage.

Closes #19352