Test Suite: Print previous exceptions
Since Ruby 2.1, when an unhandled exceptions is printed, any previous
exceptions at the time the exception was raised are also printed. That's
useful to for debugging to figure out the exact cause of an exception.

Cucumber does not print previous exceptions, so we extend it's
print_exception method to also print any previous exceptions.