Revert "Test Suite: Use Dogtail for some Greeter steps"
The test suite changes of this commit require the updates to
tails-autotest-remote-shell from commit
"Test Suite: Fix Dogtail not working when running as Debian-gdm user"
to work. That causes some scenarios of the
"Upgrading an old Tails USB installation" feature to fail, because the
old version of Tails used by that test does not have the updates to
tails-autotest-remote-shell yet.

To avoid that, lets wait until the next release and then revert this
commit again.