Dump gnome-shell env as amnesia
There are scenarios in which amnesia can control the gnome-shell
environment which is dumped at boot, so we have to ensure that
only processes running as amnesia (or a less privileged user) export it
in their environment. Since we have to do that, we can just store the
env file as amnesia and drop the part where we copy the file as root.