refactor: Rename gnome-env -> user-env
The environment we're dumping doesn't actually have anything to do with
GNOME or GNOME Shell, it's the default environment of a user unit
started by systemd plus some display related variables set by us in
dump-user-env. The name user-env therefore better reflects its content.

Also use the name exec-with-user-env instead of exec-in-user-env to make
it more clear that we're not replacing the current environment.