dump-user-env: Only dump specific environment variables
Dumping all environment variables caused failure in the test suite where
we use the environment dumped by amnesia in a root process. That caused
the PATH variable to not include /sbin anymore.

The environment variables we dump here are the same that were dumped in
the old implementation in gnome.sh.