Test suite: Translate Dogtail args to $lang_code
We used to just call `gettext` in the VM, which uses the current system
language. Now we always translate to the language specified in
$language, which allows us to also translate strings to the language
configured in the greeter.

Note that this required changing the lang_code of chinese from zh to
zh_CN because otherwise `gettext` wouldn't translate it. This should not
cause anything to break because it apparently for Chinese, the lang_code
isn't used anywhere else currently.