Test suite: Retry restoring snapshot 3 times
Pressing escape repeatedly did not fix #18953 yet, but the snapshots
for which the error occurred were restored successfully multiple times
in other tests from the same test run, which makes it likely that
trying again to restore it could succeed.

In the cases where restoring the snapshot succeeded, the display was
activated within one second, so we reduce the time we try to wake up the
display to 10 seconds.

Closes #18953