Test suite: Avoid TCA scenario failing if the error panel is briefly shown
There is race condition in TCA which sometimes causes it to briefly show
the error panel before it shows the "successfully connected" panel. We
decided to work around that in the test suite, so we now only let the
test fail if this panel is visible for more than 3 seconds.

Closes #19591