udisks2: Use LUKS2 by default
This checks in a udisks2.conf config file with the (current) default
values except that the "encryption" option is set to "luks2" instead of
"luks1".

This is not only relevant for creating the Persistent Storage partition
(we could let udisks create a LUKS2 partition there without changing the
config file) but also for other encrypted partitions which the user
might create via udisks / GNOME Disks.

Closes #15450