tps: Don't restrict CPU usage during Persistent Storage creation
Restricting the CPU usage during Persistent Storage creation restricts
the resources available during the benchmark which cryptsetup luksFormat
performs to figure out the argon2id parameters. In a test, this resulted
in a weaker memory parameter being chosen (563924 instead of 1005154).

I couldn't find a way to configure the systemd service to use all the
CPU cores and just leave 20% of one core - that should be enough to keep
the system responsive and still give enough resources to the benchmark.

As a workaround, we now make it clear to the user in the job status
message that the system might become unresponsive for a few seconds.