tps: Reduce the number of triggered udev events
We were triggering udev events for all devices, which spammed the
journal with messages from upowerd and udisksd and is quite inefficient.

We introduced calls to `udevadm trigger` because we saw it happen that,
after an operation which changed a device, a subsequent udisks operation
which uses that device fails because the device is not accessible to
udisks yet.

This commit removes all calls to `udevadm trigger` after synchronous
calls to udisks. The assumption is that the synchronous udisks call
should only return once the changed device is available to udisks.

We still call `udevadm trigger` after a non-udisks operation which
changes a device that is used by a subsequent udisks operation - but we
specify the exact device which we want udev events triggered for.

Lets see if we run into race conditions again with this change.