Allow initramfs size larger than 32 MiB
In #17320, some MacBook models ran into a kernel panic during boot when
the initramfs size was larger than 32 MiB, so we have since made the
effort to keep the size below that limit. That turned out to be
increasingly difficult, requiring us to manually figure out binaries
(mostly drivers) which we could safely delete from the initramfs.

In #19663, sajolida tested with one of the affected models, a
MacBook Pro 8.1, that the kernel panic does not happen anymore with the
latest Tails version, so we assume that the issue has been fixed by some
update.

Closes #19663