tca: Don't read user-env file as root
This reduces code duplication by re-using run_in_netns() in tca. To
achieve that, the code which decices if an error message should be
displayed instead of running tca is split off into a separate
executable. That also required making the tails.network config file
readable by the amnesia user.

Closes #19677