tails-installer: Let user choose passphrase for clone Persistent Storage
We don't have to force the user to use the same passphrase as for the
old Persistent Storage. This provides a less surprising UX and allows
the user to choose a better passphrase than they had before (at the cost
that they might confuse the passphrases of the two Tails devices, so IMO
the recommended method should be to change the passphrase of the old
device first and then clone with the same passphrase).