Test suite: actually "simulate" a failing connection during time sync
We want the "Fix clock" button to appear during the Tor connection
failure, but it actually doesn't for a 404, it needs an real
connection failure. This test used to (incorrectly) work any way prior
to enabling Dogtail's showingOnly by default (refs: tails/tails#19717)
since we then could find the "Fix clock" button even if it isn't
showing, and even press it using the a11y "press" action. But now that
showingOnly is enabled we need the test to actually be correct.