Test suite: deal with test relying on showingOnly not being enabled
With Dogtail's showingOnly disabled it would find the check box even
if it was not visible. Luckily we further check that the check box is
sensitive, so this test was correct all along.

Note that this commit also could have been simply adding `showingOnly:
false` to the `child('Clone...)` call, but this felt more Ruby:ish.

refs: tails/tails#19717