Test suite: deal with test relying on showingOnly not being enabled
Thunderbird's AppMenu button is only visible in the main view, but we
didn't fully close the account setup wizard, so with Dogtail's
showingOnly now enabled (tails/tails#19717) the AppMenu could not be
found. I suspect that it was sufficient to close the wizard with just
the Cancel but at some point the confirmation prompt was added.