TCA: Fix error not displayed after QR code scanning failure
In such cases, TCA would raise an exception while trying to display the error to
the user: "TypeError: 'dict' object is not callable" for "_"

This is a regression introduced in 5.14 by 31e4b9df. One instance of the bug was
fixed in 30b6842e, but some others remained, or could start happening whenever
we start using "_" to localize strings in the affected code paths.