Test suite: embed old image next to match in automatic image bumping mode
If you wonder why this commit is on a branch that doesn't bump images,
the explanation is simple: there we're tons of bumps, which wasn't
surprising for a Tor Browser major upgrade, but that was before
noticing the "new fontconfig path" and fixing that, which made all
bumps moot. Oh well.