Browsers: deal with another extension directory from TBB 12.5 (refs: tails/tails#19724)
In 12.5 Browser/distribution/extensions was added and contain a copy
of the only extension in the profile's extensions directory, namely
NoScript, and if we leave it it will be enabled in the Unsafe
Browser.

The added mv command will fail and abort the build if
Browser/distribution/extensions is removed or empty, so we'll notice
if this (probable?) oversight is corrected.