onion-grater: only use address comparison for event ownership
PID-based matching is racy and inherently insecure. In fact, it isn't
clear to me at the moment when the address comparison wouldn't be
sufficient. In any case, "event ownership" (and the code relying on
that) is only relevant for filters that have restrict-stream-events
set, which is only Tor Browser, for which self.client_pid will be set
to None (so the removed check could never return True) since it uses
netns-based filter matching, so address comparison clearly is
sufficient in this only case we have to support.